refactor(types): import domain types from core instead of re-exports
ADR-0004 asks importers to name the file where a symbol is defined, and
twelve route-local components were reaching Product and PremiumFeature
through services instead.

The audit read this as services owning the types, but they were already
defined in core/domain: product.ts exports the Product alias and
premium-feature.ts the PremiumFeature interface. services/product.ts
duplicated the alias and services/premiumFeature.ts re-exported it, so
nothing has to move — only the paths change, and the symbol names stay.

Drop both re-exports, take ProductJSON from core inside the service, and
remove the model's re-export line, which no importer used.

The point is not tidiness: UI files no longer name a module that carries
import "server-only". Dropping the type keyword by accident would have
pulled a server module into the client bundle, and that path is now
structurally gone.
