+++ src/app/(admin)/admin/products/new/_containers/ProductRegistrationForm.tsx
@@ -6,7 +6,7 @@ import { toast } from "sonner";
 
 import { createProduct } from "@/actions/createProduct";
 import type { APIResponse } from "@/core/domain/error";
-import type { PremiumFeature } from "@/services/premiumFeature";
+import type { PremiumFeature } from "@/core/domain/premium-feature";
 import { ProductRegistrationForm as PureProductRegistrationForm } from "../_components/ProductRegistrationForm";
 import { routes } from "@/core/domain/routes";
 export function ProductRegistrationForm({
