+++ src/app/(main)/(products)/products/[category]/_containers/ProductCatalog.tsx
@@ -3,7 +3,7 @@
 import { useProducts } from "@/ui/hooks/useProducts";
 import { usePremiumFeature } from "@/ui/hooks/usePremiumFeatures";
 import { ProductCatalog as ProductCatalogView } from "@/app/(main)/(products)/products/[category]/_components/ProductCatalog";
-import type { Product } from "@/services/product";
+import type { Product } from "@/core/domain/product";
 import type { ProductCategory, SubCategory } from "@/core/domain/product-category";
 
 export function ProductCatalog({
