+++ src/app/(main)/page.tsx
@@ -1,7 +1,7 @@
 export const dynamic = "force-dynamic";
 
 import { HomeTemplate } from "@/app/(main)/_components/HomeTemplate";
-import type { Product } from "@/services/product";
+import type { Product } from "@/core/domain/product";
 import { getAvailableSubCategoriesService, getPopularProductsService } from "@/services/product";
 import type { AvailableSubCategory } from "@/core/domain/product-category";
 import { POPULAR_PRODUCTS_LIMIT } from "@/core/domain/product";
