+++ src/services/premiumFeature.ts
@@ -4,7 +4,6 @@ import { FeatureModel } from "@/models/feature.model";
 import type { PremiumFeatureDto } from "@/core/schemas/request/premiumFeature.schema";
 import type { PremiumFeature } from "@/core/domain/premium-feature";
 import { AppError } from "@/core/domain/error";
-export type { PremiumFeature } from "@/core/domain/premium-feature";
 import { dbConnect } from "@/db/connect";
 
 import mongoose from "mongoose";
